+++ packages/breadcrumb/src/index.ts
@@ -1,1 +1,7 @@
-export * from "./breadcrumb"
+export { Breadcrumb, type BreadcrumbProps } from "./breadcrumb"
+export { BreadcrumbItem, type BreadcrumbItemProps } from "./breadcrumb-item"
+export {
+  BreadcrumbSeparator,
+  type BreadcrumbSeparatorProps,
+} from "./breadcrumb-separator"
+export { BreadcrumbLink, type BreadcrumbLinkProps } from "./breadcrumb-link"
