+++ packages/breadcrumb/stories/breadcrumb.stories.tsx
@@ -1,5 +1,4 @@
 import { ChevronRightIcon } from "@chakra-ui/icons"
-// import { extendTheme, ThemeProvider, useTheme } from "@chakra-ui/react"
 import * as React from "react"
 import { BrowserRouter, Link } from "react-router-dom"
 import { Breadcrumb, BreadcrumbItem, BreadcrumbLink } from "../src"
@@ -8,7 +7,7 @@ export default {
   title: "Components / Navigation / Breadcrumb",
 }
 
-export const Default = () => (
+export const Basic = () => (
   <BrowserRouter>
     <Breadcrumb spacing="4">
       <BreadcrumbItem>
@@ -28,7 +27,7 @@ export const Default = () => (
   </BrowserRouter>
 )
 
-export const Separator = () => (
+export const WithSeparator = () => (
   <Breadcrumb separator=">">
     <BreadcrumbItem>
       <BreadcrumbLink href="#">Home</BreadcrumbLink>
@@ -44,7 +43,7 @@ export const Separator = () => (
   </Breadcrumb>
 )
 
-export const SeparatorV2 = () => (
+export const WithCustomSeparator = () => (
   <Breadcrumb spacing="8px" separator={<ChevronRightIcon color="gray.300" />}>
     <BreadcrumbItem>
       <BreadcrumbLink href="/">Home</BreadcrumbLink>
@@ -59,58 +58,3 @@ export const SeparatorV2 = () => (
     </BreadcrumbItem>
   </Breadcrumb>
 )
-
-// export const WithThemeOverrides = () => {
-//   const currentTheme = useTheme()
-//   const theme = extendTheme(
-//     {
-//       components: {
-//         Breadcrumb: {
-//           baseStyle: {
-//             container: {
-//               borderWidth: 4,
-//               padding: 2,
-//             },
-//             item: {
-//               borderWidth: 2,
-//               borderRadius: "full",
-//               textTransform: "uppercase",
-//             },
-//             link: {
-//               color: "red.500",
-//               "&[aria-current=page]": {
-//                 color: "blue.500",
-//               },
-//             },
-//             separator: {
-//               borderWidth: 4,
-//               borderColor: "red.300",
-//             },
-//           },
-//         },
-//       },
-//     },
-//     currentTheme,
-//   )
-
-//   return (
-//     <ThemeProvider theme={theme}>
-//       <Breadcrumb
-//         spacing="8px"
-//         separator={<ChevronRightIcon color="gray.300" />}
-//       >
-//         <BreadcrumbItem>
-//           <BreadcrumbLink href="/">Home</BreadcrumbLink>
-//         </BreadcrumbItem>
-
-//         <BreadcrumbItem>
-//           <BreadcrumbLink href="/about">About</BreadcrumbLink>
-//         </BreadcrumbItem>
-
-//         <BreadcrumbItem isCurrentPage>
-//           <BreadcrumbLink href="/contact">Contact</BreadcrumbLink>
-//         </BreadcrumbItem>
-//       </Breadcrumb>
-//     </ThemeProvider>
-//   )
-// }
